+++ ExposurePlayback/Services/ExposureAnalytics.swift
@@ -72,7 +72,8 @@ public class ExposureAnalytics {
     internal enum ExternalPlayback {
         case none
         case chromecast
-        case airplay
+        case endedByActivatingAirplay
+        case startedAsAirplay
     }
     /// Tracks if external playback was started
     internal var externalPlayback: ExternalPlayback = .none
@@ -87,7 +88,7 @@ public class ExposureAnalytics {
     
     /// Instruct the analytics engine that the player is transitioning from local playback to `Airplay`
     public func startedAirplay() {
-        externalPlayback = .airplay
+        externalPlayback = .endedByActivatingAirplay
     }
 }
 
@@ -122,6 +123,7 @@ extension ExposureAnalytics {
     }
 }
 
+import AVFoundation
 extension ExposureAnalytics: ExposureStreamingAnalyticsProvider {
     private func autoplay<Tech>(tech: Tech) -> Bool {
         if let tech = tech as? MediaPlayback {
@@ -138,8 +140,12 @@ extension ExposureAnalytics: ExposureStreamingAnalyticsProvider {
                                        assetData: PlaybackIdentifier.from(playable: playable),
                                        autoPlay: autoplay(tech: tech))
         
+        /// EMP-11647: If this is an Airplay session, set `Playback.Device.Info.type = AirPlay`
+        let connectedAirplayPorts = AVAudioSession.sharedInstance().currentRoute.outputs.filter{ $0.portType == AVAudioSessionPortAirPlay }
+        let isAirplaySession = !connectedAirplayPorts.isEmpty ? AVAudioSessionPortAirPlay : nil
+        
         /// 2. DeviceInfo
-        let deviceInfo = DeviceInfo(timestamp: Date().millisecondsSince1970)
+        let deviceInfo = DeviceInfo(timestamp: Date().millisecondsSince1970, type: isAirplaySession)
         
         /// 3. Store startup events
         var current = startupEvents
@@ -171,6 +177,11 @@ extension ExposureAnalytics: ExposureStreamingAnalyticsProvider {
     /// - parameter heartbeatsProvider: Will deliver heartbeats metadata during the session
     public func finalizePreparation<Tech, Source>(tech: Tech, source: Source, playSessionId: String, heartbeatsProvider: @escaping () -> AnalyticsEvent?) where Tech : PlaybackTech, Source : MediaSource {
         let events = startupEvents
+        
+        if let tech = tech as? HLSNative<ExposureContext>, tech.isExternalPlaybackActive {
+            /// Session was started with Airplay active. This session should be terminated with ´Playback.StopAirplay`
+            externalPlayback = .startedAsAirplay
+        }
         
         dispatcher = Dispatcher(environment: environment,
                                 sessionToken: sessionToken,
@@ -267,9 +278,13 @@ extension ExposureAnalytics: AnalyticsProvider {
         }
         
         switch externalPlayback {
-        case .airplay:
+        case .endedByActivatingAirplay:
             let event = Playback.StartAirplay(timestamp: Date().millisecondsSince1970,
                                               offsetTime: offsetTime(for: source, using: tech))
+            dispatcher?.enqueue(event: event)
+        case .startedAsAirplay:
+            let event = Playback.StopAirplay(timestamp: Date().millisecondsSince1970,
+                                             offsetTime: offsetTime(for: source, using: tech))
             dispatcher?.enqueue(event: event)
         case .chromecast:
             let event = Playback.StartCasting(timestamp: Date().millisecondsSince1970,